fix: resolve circular dependency between navigation hooks
- Add useRouteDetection hook to separate route detection logic
- Refactor useNavigation to use useRouteDetection instead of direct location usage
- Update useLastDeck to use useRouteDetection instead of useNavigation
- Break circular dependency: useNavigation -> useLastDeck -> useNavigation

Root Cause:
- useNavigation imported useLastDeck
- useLastDeck imported useNavigation
- This created infinite recursion when React tried to resolve hooks

Solution:
- Extract route detection logic into independent useRouteDetection hook
- Both useNavigation and useLastDeck now use useRouteDetection
- No more circular dependencies, clean separation of concerns

This fixes the 'Maximum call stack size exceeded' error that was preventing the app from loading properly.
